backend/requirements.txt
pandas
vectorbt
yfinance
scikit-learn
polygon-api-client
dotenv
joblib
flask
flask-socketio
flask-cors
python-socketio